refactor: move rcp-receipt types to rpc crate
