+++ crates/core/types/block.rs
@@ -1,6 +1,6 @@
 use super::{
-    ReceiptBlockInfo, BASE_FEE_MAX_CHANGE_DENOMINATOR, ELASTICITY_MULTIPLIER,
-    GAS_LIMIT_ADJUSTMENT_FACTOR, GAS_LIMIT_MINIMUM,
+    BASE_FEE_MAX_CHANGE_DENOMINATOR, ELASTICITY_MULTIPLIER, GAS_LIMIT_ADJUSTMENT_FACTOR,
+    GAS_LIMIT_MINIMUM,
 };
 use crate::{
     rlp::{
@@ -270,13 +270,6 @@ impl BlockHeader {
         let mut buf = vec![];
         self.encode(&mut buf);
         keccak(buf)
-    }
-
-    pub fn receipt_info(&self) -> ReceiptBlockInfo {
-        ReceiptBlockInfo {
-            block_hash: self.compute_block_hash(),
-            block_number: self.number,
-        }
     }
 }
