+++ crates/core/types/receipt.rs
@@ -6,10 +6,9 @@ use crate::rlp::{
 };
 use bytes::Bytes;
 use ethereum_types::{Address, Bloom, H256};
-use revm::primitives::Address as EvmAddress;
 use serde::Serialize;
 
-use super::{BlockHash, BlockNumber, TxKind, TxType};
+use super::TxType;
 pub type Index = u64;
 
 /// Result of a transaction
@@ -101,10 +100,10 @@ impl RLPDecode for Receipt {
 /// Data record produced during the execution of a transaction.
 #[derive(Clone, Debug, PartialEq, Eq, Serialize)]
 pub struct Log {
-    address: Address,
-    topics: Vec<H256>,
+    pub address: Address,
+    pub topics: Vec<H256>,
     #[serde(with = "crate::serde_utils::bytes")]
-    data: Bytes,
+    pub data: Bytes,
 }
 
 impl RLPEncode for Log {
@@ -129,110 +128,5 @@ impl RLPDecode for Log {
             data,
         };
         Ok((log, decoder.finish()?))
-    }
-}
-
-// Struct used by RPC
-#[derive(Debug, Serialize)]
-pub struct ReceiptWithTxAndBlockInfo {
-    #[serde(flatten)]
-    pub receipt: Receipt,
-    #[serde(flatten)]
-    pub tx_info: ReceiptTxInfo,
-    #[serde(flatten)]
-    pub block_info: ReceiptBlockInfo,
-}
-
-#[derive(Debug, Serialize, Clone)]
-#[serde(rename_all = "camelCase")]
-pub struct ReceiptBlockInfo {
-    pub block_hash: BlockHash,
-    #[serde(with = "crate::serde_utils::u64::hex_str")]
-    pub block_number: BlockNumber,
-}
-
-#[derive(Debug, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ReceiptTxInfo {
-    pub transaction_hash: H256,
-    #[serde(with = "crate::serde_utils::u64::hex_str")]
-    pub transaction_index: u64,
-    pub from: Address,
-    pub to: Option<Address>,
-    pub contract_address: Option<Address>,
-    #[serde(with = "crate::serde_utils::u64::hex_str")]
-    pub gas_used: u64,
-    #[serde(with = "crate::serde_utils::u64::hex_str")]
-    pub effective_gas_price: u64,
-}
-
-impl ReceiptTxInfo {
-    pub fn new(
-        transaction_hash: H256,
-        transaction_index: u64,
-        nonce: u64,
-        from: Address,
-        tx_kind: TxKind,
-        gas_used: u64,
-        effective_gas_price: u64,
-    ) -> ReceiptTxInfo {
-        let (contract_address, to) = match tx_kind {
-            TxKind::Create => (
-                Some(Address::from_slice(
-                    EvmAddress(from.0.into()).create(nonce).0.as_ref(),
-                )),
-                None,
-            ),
-            TxKind::Call(addr) => (None, Some(addr)),
-        };
-        Self {
-            transaction_hash,
-            transaction_index,
-            from,
-            to,
-            contract_address,
-            gas_used,
-            effective_gas_price,
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use hex_literal::hex;
-
-    #[test]
-    fn serialize_receipt() {
-        let receipt = ReceiptWithTxAndBlockInfo {
-            receipt: Receipt {
-                tx_type: TxType::EIP4844,
-                succeeded: true,
-                cumulative_gas_used: 147,
-                bloom: Bloom::zero(),
-                logs: vec![Log {
-                    address: Address::zero(),
-                    topics: vec![],
-                    data: Bytes::from_static(b"strawberry"),
-                }],
-            },
-            tx_info: ReceiptTxInfo {
-                transaction_hash: H256::zero(),
-                transaction_index: 1,
-                from: Address::zero(),
-                to: Some(Address::from(hex!(
-                    "7435ed30a8b4aeb0877cef0c6e8cffe834eb865f"
-                ))),
-                contract_address: None,
-                gas_used: 147,
-                effective_gas_price: 157,
-            },
-            block_info: ReceiptBlockInfo {
-                block_hash: BlockHash::zero(),
-                block_number: 3,
-            },
-        };
-        let expected = r#"{"type":"0x3","status":"0x1","cumulativeGasUsed":"0x93","logsBloom":"0x00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000","logs":[{"address":"0x0000000000000000000000000000000000000000","topics":[],"data":"0x73747261776265727279"}],"transactionHash":"0x0000000000000000000000000000000000000000000000000000000000000000","transactionIndex":"0x1","from":"0x0000000000000000000000000000000000000000","to":"0x7435ed30a8b4aeb0877cef0c6e8cffe834eb865f","contractAddress":null,"gasUsed":"0x93","effectiveGasPrice":"0x9d","blockHash":"0x0000000000000000000000000000000000000000000000000000000000000000","blockNumber":"0x3"}"#;
-        assert_eq!(serde_json::to_string(&receipt).unwrap(), expected);
     }
 }
