+++ crates/rpc/Cargo.toml
@@ -21,5 +21,8 @@ hex.workspace = true
 axum-extra = {version = "0.9.3", features = ["typed-header"]}
 jsonwebtoken = "9.3.0"
 
+[dev-dependencies]
+hex-literal = "0.4.1"
+
 [lib]
 path = "./rpc.rs"
